test: move test utils to tests folder
